requirements.txt
# core
groq>=0.9.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0

# langgraph + langchain
langchain>=0.2.0
langchain-groq>=0.1.6
langchain-core>=0.2.0
langgraph>=0.1.0
